serial revolution

+++ Project_wsl_2026/Software/main_tennsy4.0/lib/Camera/serial/serial.hpp
@@ -13,24 +13,24 @@ void Camera_update();
 bool is_field_exist();
 
 // コートの角度を出力
-int16_t get_field_deg();
+int get_field_deg();
 
 // 黄色ゴールが存在するかどうか
 bool is_yellow_goal_exist();
 
 // 黄色ゴールの角度を出力
-int16_t get_yellow_goal_deg();
+int get_yellow_goal_deg();
 
 // 黄色ゴールの距離を出力
-int16_t get_yellow_goal_distance();
+int get_yellow_goal_distance();
 
 // 青ゴールが存在するかどうか
 bool is_blue_goal_exist();
 
 // 青色ゴールの角度を出力
-int16_t get_blue_goal_deg();
+int get_blue_goal_deg();
 
 // 青色ゴールの距離を出力
-int16_t get_blue_goal_distance();
+int get_blue_goal_distance();
 
 #endif
